+++ front/app/containers/Admin/tools/PublicAPI/apIllustration.png
