+++ front/app/containers/Admin/tools/PublicAPI/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Box, Text, Badge, colors } from '@citizenlab/cl2-component-library';
 import Button from 'components/UI/Button';
 import { useIntl } from 'utils/cl-intl';
-import messages from './messages';
+import messages from '../messages';
 import apiImage from './apIllustration.png';
 
 export const PublicAPI = () => {
